requirements.txt
streamlit
pandas
yfinance
torch
networkx
pyvis
plotly
python-dotenv
requests
numpy
scikit-learn
newsapi-python